connect property and move client variable to clobj
