+++ src/message_broker/client/client.rs
@@ -1,5 +1,5 @@
-use std::{fmt::Display, io, net::SocketAddr};
-use tokio::{io::{AsyncReadExt, AsyncWriteExt, ReadHalf, WriteHalf}, net::TcpStream};
+use std::io;
+use tokio::{io::{AsyncReadExt, AsyncWriteExt}, net::TcpStream};
 use crate::{
     connection::{
         handshake::MqttConnectedResponse, 
@@ -10,164 +10,45 @@ use crate::{
     helper::time::sys_now, 
     protocol::v5::connack::ConnackPacket
 };
-use super::{storage::ClientStore, SessionController};
-
-extern crate tokio;
-
-#[derive(Debug, Eq, Ord, Clone)]
-pub struct ClientID {
-    id: String,
-    hash: u32,
-}
-
-impl ClientID {
-    pub fn new(raw_clid: String) -> ClientID {
-        Self{
-            hash: Self::hash(&raw_clid),
-            id: raw_clid,
-        }
-    }
-
-    /// murmurhash 3
-    fn hash(raw_clid: &str) -> u32 {
-        const C1: u32 = 0xcc9e2d51;
-        const C2: u32 = 0x1b873593;
-        const SEED: u32 = 0;
-        
-        let mut hash = SEED;
-        let data = raw_clid.as_bytes();
-
-        let nblocks = data.len() / 4;
-
-        for i in 0..nblocks {
-            let mut k = u32::from_le_bytes([data[4 * i], data[4 * i + 1], data[4 * i + 2], data[4 * i + 3]]);
-            k = k.wrapping_mul(C1);
-            k = k.rotate_left(15);
-            k = k.wrapping_mul(C2);
-            
-            hash ^= k;
-            hash = hash.rotate_left(13);
-            hash = hash.wrapping_mul(5).wrapping_add(0xe6546b64);
-        }
-
-        let tail = &data[nblocks * 4..];
-        let mut k1 = 0;
-        match tail.len() {
-            3 => k1 ^= (tail[2] as u32) << 16,
-            2 => k1 ^= (tail[1] as u32) << 8,
-            1 => k1 ^= tail[0] as u32,
-            _ => (),
-        }
-        k1 = k1.wrapping_mul(C1);
-        k1 = k1.rotate_left(15);
-        k1 = k1.wrapping_mul(C2);
-        hash ^= k1;
-
-        hash ^= data.len() as u32;
-        hash ^= hash >> 16;
-        hash = hash.wrapping_mul(0x85ebca6b);
-        hash ^= hash >> 13;
-        hash = hash.wrapping_mul(0xc2b2ae35);
-        hash ^= hash >> 16;
-
-        hash
-    }
-}
-
-impl Display for ClientID {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.id)
-    }
-}
-
-impl PartialEq for ClientID {
-    fn eq(&self, other: &Self) -> bool {
-        if self.id.len() != other.id.len() {
-            return false;
-        }
-        
-        self.hash == other.hash
-    }
-
-    fn ne(&self, other: &Self) -> bool {
-        if self.id.len() == other.id.len() {
-            return false;
-        }
-
-        self.hash != other.hash
-    }
-}
-
-impl PartialOrd for ClientID {
-    fn ge(&self, other: &Self) -> bool {
-        self.hash >= other.hash
-    }
-
-    fn gt(&self, other: &Self) -> bool {
-        self.hash > other.hash
-    }
-
-    fn le(&self, other: &Self) -> bool {
-        self.hash <= other.hash
-    }
-
-    fn lt(&self, other: &Self) -> bool {
-        self.hash < other.hash
-    }
-
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.hash.cmp(&other.hash))
-    }
-}
-
-enum SocketInner<S, P> {
-    Secured(S),
-    Plain(P)
-}
-
-pub struct Socket<S, P> {
-    r: SocketInner<ReadHalf<S>, ReadHalf<P>>,
-    w: SocketInner<WriteHalf<S>, WriteHalf<P>>
-}
-
-impl Socket<SecuredStream, TcpStream> {
-    fn new(socket: SocketConnection) -> Self {
-        match socket {
-            SocketConnection::Plain(p) => {
-                let (r, w) = tokio::io::split(p);
-                Socket {
-                    r: SocketInner::Plain(r),
-                    w: SocketInner::Plain(w),
-                }
-            },
-            SocketConnection::Secure(s) => {
-                let (r, w) = tokio::io::split(s);
-                Socket {
-                    r: SocketInner::Secured(r),
-                    w: SocketInner::Secured(w),
-                }
-            }
-        }
-    }
-}
+use super::{clobj::{ClientID, Socket, SocketInner}, storage::{ClientStore, MetaData}, SessionController};
 
 
 #[allow(dead_code)]
 pub struct Client {
     pub(super) conid: ConnectionID,
     pub clid: ClientID,
-    addr: SocketAddr,
     pub socket: Socket<SecuredStream, TcpStream>,
     protocol_level: u8,
+    pub session: Session,
+    pub limit: Limiter,
+    pub storage: ClientStore
+}
+
+pub struct Session {
     ttl: u64,
     keep_alive: u16,
     expr_interval: u32,
-    pub storage: ClientStore
+}
+
+#[derive(Default)]
+pub struct Limiter {
+    receive_maximum: Option<u16>,
+    maximum_packet_size: Option<u16>,
+    topic_alias_maximum: Option<u16>,
+}
+
+impl Limiter {
+    pub fn new(
+        receive_maximum: Option<u16>, 
+        maximum_packet_size: Option<u16>, 
+        topic_alias_maximum: Option<u16>
+    ) -> Self {
+        Self { maximum_packet_size, receive_maximum, topic_alias_maximum }
+    }
 }
 
 pub struct UpdateClient {
     pub conid: Option<ConnectionID>,
-    pub addr: Option<SocketAddr>,
     pub socket: Option<SocketConnection>,
     pub protocol_level: Option<u8>,
     pub keep_alive: Option<u16>,
@@ -178,23 +59,38 @@ impl Client {
     pub async fn new(
         conid: ConnectionID,
         socket: SocketConnection,
-        addr: SocketAddr,
         clid: ClientID,
         keep_alive: u16,
         expr_interval: u32,
-        protocol_level: u8
+        protocol_level: u8,
+        limit: Limiter
     ) -> Self {
         let socket = Socket::new(socket);
+        let ttl = sys_now() + keep_alive as u64;
         let keep_alive = keep_alive.max(60);
-        let storage = ClientStore::new(&clid).await.unwrap();
-        Self {
-            conid,
-            addr,
-            socket,
-            clid,
-            ttl: sys_now() + keep_alive as u64,
+        let mdata = MetaData {
+            expr_interval,
+            keep_alive_interval: keep_alive,
+            protocol_level,
+            maximum_packet_size: limit.maximum_packet_size.unwrap_or_default(),
+            receive_maximum: limit.receive_maximum.unwrap_or_default(),
+            topic_alias_maximum: limit.topic_alias_maximum.unwrap_or_default(),
+            user_properties: Vec::new()
+        };
+
+        let session =  Session {
             expr_interval,
             keep_alive,
+            ttl
+        };
+
+        let storage = ClientStore::new(&clid, &mdata, None).await.unwrap();
+        Self {
+            conid,
+            socket,
+            clid,
+            session,
+            limit,
             protocol_level,
             storage
         }
@@ -203,7 +99,7 @@ impl Client {
     // TODO: Error type
     pub fn restore_connection(&mut self, bucket: &mut UpdateClient) -> io::Result<()> {
         let now = sys_now();
-        if !self.is_alive(now) && self.is_expired(now) {
+        if !self.session.is_alive(now) && self.session.is_expired(now) {
            return Err(io::Error::new(io::ErrorKind::Other, String::from("connection is expired"))); 
         }
 
@@ -213,11 +109,9 @@ impl Client {
         };
 
         if let Some(keep_alive) = bucket.keep_alive.take() {
-            self.keep_alive = keep_alive;
+            self.session.keep_alive = keep_alive;
         }
-        if let  Some(addr) = bucket.addr.take() {
-            self.addr = addr
-        }
+
         if let Some(connid) = bucket.conid.take() {
             self.conid = connid;
         }
@@ -229,7 +123,7 @@ impl Client {
     }
 }
 
-impl SessionController for Client {
+impl SessionController for Session {
     fn is_alive(&self, t: u64) -> bool {
         self.ttl >= t
     }
