+++ src/message_broker/msg_state.rs
@@ -1,6 +1,8 @@
 use std::sync::Arc;
 use tokio::sync::RwLock;
-use crate::{helper::time::sys_now, message_broker::client::client::ClientID};
+use crate::helper::time::sys_now;
+
+use super::client::clobj::ClientID;
 
 #[derive(Clone)]
 pub struct MessageCoordinator {
